Problem: 'basic' module is still pretty big
Solution: extract binaries related functionality into its own module
