+++ src/script/mod.rs
@@ -349,6 +349,7 @@ use pubsub;
 pub mod basic;
 pub mod stack;
 pub mod numbers;
+pub mod binaries;
 pub mod storage;
 pub mod timestamp_hlc;
 pub mod hash;
@@ -430,6 +431,7 @@ impl<'a> VM<'a> {
             sender: sender.clone(),
             modules: vec![Box::new(basic::Handler::new(publisher)),
                           Box::new(stack::Handler::new()),
+                          Box::new(binaries::Handler::new()),
                           Box::new(numbers::Handler::new()),
                           Box::new(storage::Handler::new(db_env, db)),
                           Box::new(hash::Handler::new()),
